Adds methods to run the Lox interperter
This commit adds two methods to run the interpreter in scripting
or interractive modes. It implements the main method delegating
to an appropriate run method based on the arguments provided.

It also adds basic error reporting that shows the line
where an error occured and a message describing the error.

It also adds ExitCode, Token and LoxError classes.