Refactors Lox class
This commit updates Lox class and makes it the responsibility of
parse and interpreter methods to check for error before performing
their actions. This is to make the code cleaner.

Moves stringify method in Interpreter class to the Util class.
Updates AstPrinter to use the stringigy method to print literals.