Adds Environment and implements variable declaration and lookup.
This commit adds a variable declaration statement to the language.
It updates the lox.bnf file adding a rule for declaration and
and variable declaration.

Updates the Parser to recognise declaration in general and
variable declaration in particular.

Updates the Interpreter to bind the name of a variable to a value
in an Environment.