Bug Fix: Correctly report the token after which an error occurs
Previously, when an error occurs after an expression or token has
been parsed, the error message displayed to the user prints the
next expression or token as the offending part instead of the one
previously parsed.

For example, when a semicolon is missing when it is required, the
offending or expression should be the previously parsed expression
instead of the one next up in the code.